auth/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
aiomysql==0.2.0
pydantic==2.5.1
pydantic-settings==2.1.0
python-jose==3.3.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.0.1
cryptography>=41.0.0
email-validator>=2.0.0
prometheus-fastapi-instrumentator>=0.1.1